refactor(provider): split BatchScheduler god-file by concern (2503->538 LOC)
Split the BatchScheduler actor core into focused same-module extension files
alongside the existing +EngineBridge/+KVEstimation/+Telemetry/+Liveness/
+KVQuantScheme: ModelLifecycle, EngineFactory, CheckpointRestore, Admission,
Submit, PrefixCacheSizing, Testing.

Method bodies are byte-identical; only access control changed. 13 members
reached across the split were widened private->internal (incl. resolving a
checkpointLayerSignatures static-vs-instance name shadow); purely-local
members stay private. Behavior unchanged.

Build green; swift test 1052/73 pass (unchanged from baseline).
